admin use only fix
Blackbox needs to be able to debug on the front-end, but only show for developers (ie. administrator role.)
Using is_admin() (a change from the original Blackbox) prevented this version from any front-end use at all.
This is a hacky fix after the fact, but it works. It could be better implemented in the Blackbox class.
And possibly provide a filter for allowed roles and/or capabilities so the plugin only loads when it is needed.